+++ packages/react-components/react-carousel-preview/src/components/Carousel/Carousel.types.ts
@@ -23,6 +23,16 @@ export type CarouselProps = ComponentProps<CarouselSlots> & {
    * Callback to notify a page change.
    */
   onValueChange?: EventHandler<CarouselValueChangeData>;
+
+  /**
+   * Circular enables the carousel to loop back around on navigation past trailing index
+   */
+  circular?: Boolean;
+
+  /**
+   * Peeking enables the next/prev carousel pages to 'peek' into the current view
+   */
+  peeking?: Boolean;
 };
 
 /**