+++ packages/react-components/react-carousel-preview/src/components/CarouselNavImageButton/useCarouselNavImageButton.ts
@@ -1,6 +1,11 @@
 import * as React from 'react';
-import { getIntrinsicElementProps, slot } from '@fluentui/react-utilities';
+import { getIntrinsicElementProps, isHTMLElement, slot, useEventCallback } from '@fluentui/react-utilities';
 import type { CarouselNavImageButtonProps, CarouselNavImageButtonState } from './CarouselNavImageButton.types';
+import { ARIAButtonElement, ARIAButtonSlotProps, useARIAButtonProps } from '@fluentui/react-aria';
+import { useTabsterAttributes } from '@fluentui/react-tabster';
+import { useCarouselContext_unstable } from '../CarouselContext';
+import { useCarouselNavContext } from '../CarouselNav/CarouselNavContext';
+import { useCarouselStore_unstable } from '../useCarouselStore';
 
 /**
  * Create the state required to render CarouselNavImageButton.
@@ -9,26 +14,64 @@ import type { CarouselNavImageButtonProps, CarouselNavImageButtonState } from '.
  * before being passed to renderCarouselNavImageButton_unstable.
  *
  * @param props - props from this instance of CarouselNavImageButton
- * @param ref - reference to root HTMLDivElement of CarouselNavImageButton
+ * @param ref - reference to root HTMLButtonElement | HTMLAnchorElement of CarouselNavImageButton
  */
 export const useCarouselNavImageButton_unstable = (
   props: CarouselNavImageButtonProps,
-  ref: React.Ref<HTMLDivElement>,
+  ref: React.Ref<ARIAButtonElement>,
 ): CarouselNavImageButtonState => {
-  return {
-    // TODO add appropriate props/defaults
-    components: {
-      // TODO add each slot's element type or component
-      root: 'div',
+  const { onClick, as = 'button' } = props;
+
+  const value = useCarouselNavContext();
+
+  const { selectPageByValue } = useCarouselContext_unstable();
+  const selected = useCarouselStore_unstable(snapshot => snapshot.activeValue === value);
+
+  const handleClick: ARIAButtonSlotProps['onClick'] = useEventCallback(event => {
+    onClick?.(event);
+
+    if (!event.defaultPrevented && isHTMLElement(event.target)) {
+      selectPageByValue(event, value);
+    }
+  });
+
+  const defaultTabProps = useTabsterAttributes({
+    focusable: { isDefault: selected },
+  });
+
+  const _carouselButton = slot.always<ARIAButtonSlotProps>(
+    getIntrinsicElementProps(as, useARIAButtonProps(props.as, props)),
+    {
+      elementType: 'button',
+      defaultProps: {
+        ref: ref as React.Ref<HTMLButtonElement>,
+        role: 'tab',
+        type: 'button',
+        ...defaultTabProps,
+      },
     },
-    // TODO add appropriate slots, for example:
-    // mySlot: resolveShorthand(props.mySlot),
-    root: slot.always(
-      getIntrinsicElementProps('div', {
-        ref,
-        ...props,
-      }),
-      { elementType: 'div' },
-    ),
+  );
+
+  // Override onClick
+  _carouselButton.onClick = handleClick;
+
+  const image = slot.always(
+    getIntrinsicElementProps('img', {
+      'aria-hidden': true, // Hidden as button is responsible for navigation description
+      alt: '',
+      role: 'presentation',
+      ...props.image,
+    }),
+    { elementType: 'img' },
+  );
+
+  return {
+    components: {
+      root: 'button',
+      image: 'img',
+    },
+    root: _carouselButton,
+    image,
+    selected,
   };
 };