+++ packages/react-components/react-dialog/src/index.ts
@@ -4,6 +4,7 @@ export type {
   DialogSlots,
   DialogProps,
   DialogState,
+  DialogModalType,
   DialogOpenChangeData,
   DialogOpenChangeEvent,
   DialogOpenChangeEventHandler,