+++ packages/react-components/react-migration-v0-v9/stories/GridShim/index.stories.tsx
@@ -11,45 +11,63 @@ const useStyles = makeStyles({
 });
 
 const images = [
-  <Image key="allan" fluid src="https://fabricweb.azureedge.net/fabric-website/assets/images/avatar/AllanMunger.jpg" />,
+  <Image
+    key="allan"
+    alt="allan"
+    fluid
+    src="https://fabricweb.azureedge.net/fabric-website/assets/images/avatar/AllanMunger.jpg"
+  />,
   <Image
     key="amanda"
+    alt="amanda"
     fluid
     src="https://fabricweb.azureedge.net/fabric-website/assets/images/avatar/AmandaBrady.jpg"
   />,
   <Image
     key="cameron"
+    alt="cameron"
     fluid
     src="https://fabricweb.azureedge.net/fabric-website/assets/images/avatar/CameronEvans.jpg"
   />,
   <Image
     key="carlos"
+    alt="carlos"
     fluid
     src="https://fabricweb.azureedge.net/fabric-website/assets/images/avatar/CarlosSlattery.jpg"
   />,
   <Image
     key="carole"
+    alt="carole"
     fluid
     src="https://fabricweb.azureedge.net/fabric-website/assets/images/avatar/CarolePoland.jpg"
   />,
-  <Image key="cecil" fluid src="https://fabricweb.azureedge.net/fabric-website/assets/images/avatar/CecilFolk.jpg" />,
+  <Image
+    key="cecil"
+    alt="cecil"
+    fluid
+    src="https://fabricweb.azureedge.net/fabric-website/assets/images/avatar/CecilFolk.jpg"
+  />,
   <Image
     key="celeste"
+    alt="celeste"
     fluid
     src="https://fabricweb.azureedge.net/fabric-website/assets/images/avatar/CelesteBurton.jpg"
   />,
   <Image
     key="colin"
+    alt="colin"
     fluid
     src="https://fabricweb.azureedge.net/fabric-website/assets/images/avatar/ColinBallinger.jpg"
   />,
   <Image
     key="daisy"
+    alt="daisy"
     fluid
     src="https://fabricweb.azureedge.net/fabric-website/assets/images/avatar/DaisyPhillips.jpg"
   />,
   <Image
     key="elliot"
+    alt="elliot"
     fluid
     src="https://fabricweb.azureedge.net/fabric-website/assets/images/avatar/ElliotWoodward.jpg"
   />,