+++ packages/react-components/react-motions-preview/src/factories/createPresenceComponent.test.tsx
@@ -1,4 +1,4 @@
-import { render } from '@testing-library/react';
+import { act, render } from '@testing-library/react';
 import * as React from 'react';
 
 import type { PresenceMotion } from '../types';
@@ -18,11 +18,9 @@ function createElementMock() {
   const finishMock = jest.fn();
   const animateMock = jest.fn().mockImplementation(() => ({
     cancel: jest.fn(),
+    persist: jest.fn(),
     finish: finishMock,
-    set onfinish(callback: Function) {
-      callback();
-      return;
-    },
+    finished: Promise.resolve(),
   }));
   const ElementMock = React.forwardRef<{ animate: () => void }, { onRender?: () => void }>((props, ref) => {
     React.useImperativeHandle(ref, () => ({
@@ -67,6 +65,45 @@ describe('createPresenceComponent', () => {
       );
 
       expect(animateMock).toHaveBeenCalledWith(enterKeyframes, options);
+    });
+  });
+
+  describe('onMotionFinish', () => {
+    it('is not called on first render', async () => {
+      const onMotionFinish = jest.fn();
+      const TestAtom = createPresenceComponent(motion);
+      const { ElementMock } = createElementMock();
+
+      render(
+        <TestAtom onMotionFinish={onMotionFinish}>
+          <ElementMock />
+        </TestAtom>,
+      );
+
+      expect(onMotionFinish).toHaveBeenCalledTimes(0);
+    });
+
+    it('calls "onMotionFinish" when animation finishes', async () => {
+      const onMotionFinish = jest.fn();
+      const TestAtom = createPresenceComponent(motion);
+      const { ElementMock } = createElementMock();
+
+      const { rerender } = render(
+        <TestAtom onMotionFinish={onMotionFinish} visible>
+          <ElementMock />
+        </TestAtom>,
+      );
+
+      await act(async () => {
+        rerender(
+          <TestAtom onMotionFinish={onMotionFinish} visible={false}>
+            <ElementMock />
+          </TestAtom>,
+        );
+      });
+
+      expect(onMotionFinish).toHaveBeenCalledTimes(1);
+      expect(onMotionFinish).toHaveBeenCalledWith(null, { direction: 'exit' });
     });
   });
 
@@ -116,7 +153,7 @@ describe('createPresenceComponent', () => {
   });
 
   describe('unmountOnExit', () => {
-    it('unmounts when state changes', () => {
+    it('unmounts when state changes', async () => {
       const TestAtom = createPresenceComponent(motion);
       const onRender = jest.fn();
       const { animateMock, ElementMock } = createElementMock();
@@ -134,11 +171,14 @@ describe('createPresenceComponent', () => {
       // ---
 
       jest.clearAllMocks();
-      rerender(
-        <TestAtom visible={false} unmountOnExit>
-          <ElementMock onRender={onRender} />
-        </TestAtom>,
-      );
+
+      await act(async () => {
+        rerender(
+          <TestAtom visible={false} unmountOnExit>
+            <ElementMock onRender={onRender} />
+          </TestAtom>,
+        );
+      });
 
       expect(queryByText('ElementMock')).toBe(null);
       expect(animateMock).toHaveBeenCalledWith(exitKeyframes, options);
@@ -208,7 +248,7 @@ describe('createPresenceComponent', () => {
 });
 
 describe('PresenceGroupChildContext', () => {
-  it('calls "onExit" when "visible" changes to "false"', () => {
+  it('calls "onExit" when "visible" changes to "false"', async () => {
     const onExit = jest.fn();
     const TestAtom = createPresenceComponent(motion);
     const { ElementMock } = createElementMock();
@@ -232,13 +272,15 @@ describe('PresenceGroupChildContext', () => {
 
     // ---
 
-    rerender(
-      <Wrapper visible={false}>
-        <TestAtom>
-          <ElementMock />
-        </TestAtom>
-      </Wrapper>,
-    );
+    await act(async () => {
+      rerender(
+        <Wrapper visible={false}>
+          <TestAtom>
+            <ElementMock />
+          </TestAtom>
+        </Wrapper>,
+      );
+    });
 
     expect(queryByText('ElementMock')).toBe(null);
     expect(onExit).toHaveBeenCalledTimes(1);