+++ packages/react-components/react-tag-picker/src/components/TagPickerOption/TagPickerOption.types.ts
@@ -11,9 +11,28 @@ export type TagPickerOptionSlots = Pick<OptionSlots, 'root'> & {
  * TagPickerOption Props
  */
 export type TagPickerOptionProps = ComponentProps<TagPickerOptionSlots> & {
-  children: React.ReactNode;
   value: string;
-};
+} & (
+    | {
+        /**
+         * An optional override the string value of the Option's display text,
+         * defaulting to the Option's child content.
+         * This is used as the Dropdown button's or Combobox input's value when the option is selected,
+         * and as the comparison for type-to-find keyboard functionality.
+         */
+        text?: string;
+        children: string;
+      }
+    | {
+        /**
+         * The string value of the Option's display text when the Option's children are not a string.
+         * This is used as the Dropdown button's or Combobox input's value when the option is selected,
+         * and as the comparison for type-to-find keyboard functionality.
+         */
+        text: string;
+        children?: React.ReactNode;
+      }
+  );
 
 /**
  * State used in rendering TagPickerOption