+++ packages/web-components/src/radio/index.ts
@@ -1,4 +1,5 @@
-export { Radio, RadioControl, RadioOptions } from './radio.js';
 export { definition as RadioDefinition } from './radio.definition.js';
+export { Radio } from './radio.js';
+export type { RadioControl, RadioOptions } from './radio.js';
 export { styles as RadioStyles } from './radio.styles.js';
 export { template as RadioTemplate } from './radio.template.js';