+++ packages/dom-utilities/CHANGELOG.md
@@ -1,8 +1,26 @@
 # Change Log - @fluentui/dom-utilities
 
-This log was last generated on Thu, 23 May 2024 07:28:51 GMT and should not be manually modified.
+This log was last generated on Tue, 28 May 2024 07:28:20 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [2.3.5](https://github.com/microsoft/fluentui/tree/@fluentui/dom-utilities_v2.3.5)
+
+Tue, 28 May 2024 07:28:20 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/dom-utilities_v2.3.4..@fluentui/dom-utilities_v2.3.5)
+
+### Patches
+
+- Bump @fluentui/set-version to v8.2.21 ([PR #31324](https://github.com/microsoft/fluentui/pull/31324) by beachball)
+
+## [2.3.4](https://github.com/microsoft/fluentui/tree/@fluentui/dom-utilities_v2.3.4)
+
+Fri, 24 May 2024 07:28:17 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/dom-utilities_v2.3.3..@fluentui/dom-utilities_v2.3.4)
+
+### Patches
+
+- Bump @fluentui/set-version to v8.2.20 ([commit](https://github.com/microsoft/fluentui/commit/e5b614623b9aa1ae3f1f86f3e753b934943a4601) by beachball)
 
 ## [2.3.3](https://github.com/microsoft/fluentui/tree/@fluentui/dom-utilities_v2.3.3)